fix: split FileHandleNextLine tests to separate file
Move FileHandle.nextLine tests to ProcessExtensionTests.swift to keep
FileExtensionTests.swift under the 400 line SwiftLint limit.
