+++ WhiskyKit/Tests/WhiskyKitTests/FileExtensionTests.swift
@@ -345,61 +345,3 @@ final class FileManagerReplaceDLLsTests: XCTestCase {
         XCTAssertTrue(FileManager.default.fileExists(atPath: backupURL.path(percentEncoded: false)))
     }
 }
-
-// MARK: - FileHandle.nextLine Tests
-
-final class FileHandleNextLineTests: XCTestCase {
-    var tempURL: URL!
-
-    override func setUp() {
-        super.setUp()
-        tempURL = FileManager.default.temporaryDirectory.appending(path: "nextline_\(UUID().uuidString).txt")
-    }
-
-    override func tearDown() {
-        try? FileManager.default.removeItem(at: tempURL)
-        super.tearDown()
-    }
-
-    func testNextLineWithContent() throws {
-        try Data("Hello, World!".utf8).write(to: tempURL)
-
-        let handle = try FileHandle(forReadingFrom: tempURL)
-        defer { try? handle.close() }
-
-        let line = handle.nextLine()
-        XCTAssertEqual(line, "Hello, World!")
-    }
-
-    func testNextLineWithEmptyFile() throws {
-        try Data().write(to: tempURL)
-
-        let handle = try FileHandle(forReadingFrom: tempURL)
-        defer { try? handle.close() }
-
-        let line = handle.nextLine()
-        XCTAssertNil(line)
-    }
-
-    func testNextLineWithMultipleLines() throws {
-        try Data("Line 1\nLine 2\nLine 3".utf8).write(to: tempURL)
-
-        let handle = try FileHandle(forReadingFrom: tempURL)
-        defer { try? handle.close() }
-
-        // First call gets all available data
-        let line = handle.nextLine()
-        XCTAssertNotNil(line)
-        XCTAssertTrue(line?.contains("Line 1") ?? false)
-    }
-
-    func testNextLineWithUnicodeContent() throws {
-        try Data("日本語テスト 🍷".utf8).write(to: tempURL)
-
-        let handle = try FileHandle(forReadingFrom: tempURL)
-        defer { try? handle.close() }
-
-        let line = handle.nextLine()
-        XCTAssertEqual(line, "日本語テスト 🍷")
-    }
-}
